+++ internal/display/fyne.go
@@ -149,9 +149,9 @@ func (fd *FyneDisplay) makeBackdrop() {
 	}
 }
 
-func floatToUint8(in float64) uint8 {
+func floatToUint8(in float32) uint8 {
 	// Gamma 2.0
-	in = math.Sqrt(in)
+	in = float32(math.Sqrt(float64(in)))
 	p := int(in * 255)
 	if p > 255 {
 		return 255
