+++ internal/material/material.go
@@ -9,7 +9,7 @@ import (
 
 func randomInUnitSphere(random *fastrandom.LCG) *vec3.Vec3Impl {
 	for {
-		p := vec3.Sub(vec3.ScalarMul(&vec3.Vec3Impl{X: random.Float64(), Y: random.Float64(), Z: random.Float64()}, 2.0),
+		p := vec3.Sub(vec3.ScalarMul(&vec3.Vec3Impl{X: random.Float32(), Y: random.Float32(), Z: random.Float32()}, 2.0),
 			&vec3.Vec3Impl{X: 1.0, Y: 1.0, Z: 1.0})
 		if p.SquaredLength() < 1.0 {
 			return p
@@ -22,7 +22,7 @@ func reflect(v *vec3.Vec3Impl, n *vec3.Vec3Impl) *vec3.Vec3Impl {
 	return vec3.Sub(v, vec3.ScalarMul(n, 2*vec3.Dot(v, n)))
 }
 
-func refract(v *vec3.Vec3Impl, n *vec3.Vec3Impl, niOverNt float64) (*vec3.Vec3Impl, bool) {
+func refract(v *vec3.Vec3Impl, n *vec3.Vec3Impl, niOverNt float32) (*vec3.Vec3Impl, bool) {
 	uv := vec3.UnitVector(v)
 
 	dt := vec3.Dot(uv, n)
@@ -30,14 +30,14 @@ func refract(v *vec3.Vec3Impl, n *vec3.Vec3Impl, niOverNt float64) (*vec3.Vec3Im
 	if discriminant > 0 {
 		// niOverNt * (uv - n*dt) - n*sqrt(discriminant)
 		refracted := vec3.Sub(vec3.ScalarMul(vec3.Sub(uv, vec3.ScalarMul(n, dt)), niOverNt),
-			vec3.ScalarMul(n, math.Sqrt(discriminant)))
+			vec3.ScalarMul(n, float32(math.Sqrt(float64(discriminant)))))
 		return refracted, true
 	}
 	return nil, false
 }
 
-func schlick(cosine float64, refIdx float64) float64 {
+func schlick(cosine float32, refIdx float32) float32 {
 	r0 := (1.0 - refIdx) / (1.0 + refIdx)
 	r0 = r0 * r0
-	return r0 + (1.0-r0)*math.Pow((1.0-cosine), 5)
+	return r0 + (1.0-r0)*float32(math.Pow(float64(1.0-cosine), 5))
 }
